fix: unmatched types error for system.SetProxy()

+++ internal/system/system.go
@@ -3,7 +3,7 @@
 
 package system
 
-func SetProxy(port int) error {
+func SetProxy(port uint16) error {
 	return nil
 }
